chore: refactor sort & filter func
